【api】分片上传完成后解析文件，将数据保存到redis

+++ api/pkg/api/request/app.go
@@ -4,13 +4,13 @@ import "mime/multipart"
 
 type UploadParams struct {
 	File             *multipart.FileHeader `form:"file" binding:"required"`
-	ChunkNumber      uint64                `form:"chunk_number" binding:"required,numeric"`
+	ChunkNumber      int                   `form:"chunk_number" binding:"required,numeric"`
 	ChunkSize        uint64                `form:"chunk_size" binding:"required,numeric"`
 	CurrentChunkSize uint64                `form:"current_chunk_size" binding:"required,numeric"`
 	TotalSize        uint64                `form:"total_size" binding:"required,numeric"`
 	Identifier       string                `form:"identifier" binding:"required"`
 	Filename         string                `form:"filename" binding:"required"`
-	ChunkTotal       uint32                `form:"chunk_total" binding:"required,numeric"`
+	ChunkTotal       int                   `form:"chunk_total" binding:"required,numeric"`
 }
 
 type AppInfoParams struct {
